TECH-4812: copy methods to logging_methods
